Trading/other_fx/requirements.txt
Flask==1.1.2
flask-sqlalchemy==2.4.4
numpy==1.18.1
pandas==1.2.0
scikit-learn==0.24.1
SQLAlchemy==1.3.19
xgboost==1.3.1
ta==0.7.0
lightgbm==3.1.1
psycopg2==2.8.6
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
pycaret==2.2.3
